# api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: TraderMemos API
  version: 1.0.0
  description: |
    Self-hosted trading journal API.

    Authenticate with either:
    - A short-lived session JWT from `POST /api/v1/auth/login` (web app)
    - A personal access token (`tm_pat_…`) from Settings → API or `POST /api/v1/access-tokens`

    Send the secret as `Authorization: Bearer <token>`.
servers:
  - url: /
    description: This API host
...
paths:
...
  /api/v1/me:
    get:
      tags: [Auth]
      summary: The signed-in user
      operationId: getMe
      responses:
        "200":
          description: Profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Me"
        "401":
          $ref: "#/components/responses/Error"
  /api/v1/me/password:
    put:
      tags: [Auth]
      summary: Change your own password
      description: >-
        Re-hashes after verifying the current password. Every refresh token
        minted against the old password stops working, so other devices are
        signed out at their next refresh; a fresh pair is returned so the
        caller is not. Access tokens already issued remain valid until they
        expire (15 minutes).
      operationId: changePassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [current_password, new_password]
              properties:
                current_password:
                  type: string
                new_password:
                  type: string
      responses:
        "200":
          description: New token pair for this caller
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenPair"
        "400":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
  /api/v1/me/totp/start:
    post:
      tags: [Auth]
      summary: Begin authenticator setup
      description: >-
        Mints a candidate secret and its otpauth:// URL. Nothing is stored
        until /me/totp/confirm proves a code can be read from it, so the
        client holds the candidate between the two calls.
      operationId: startTotp
      responses:
        "200":
          description: Candidate secret
          content:
            application/json:
              schema:
                type: object
                required: [secret, otpauth_url]
                properties:
                  secret:
                    type: string
                    description: Base32, for manual entry when a QR cannot be scanned.
                  otpauth_url: { type: string }
        "409":
          $ref: "#/components/responses/Error"
  /api/v1/me/totp/confirm:
    post:
      tags: [Auth]
      summary: Finish authenticator setup
      operationId: confirmTotp
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [secret, code]
              properties:
                secret: { type: string }
                code: { type: string }
      responses:
        "204": { description: Enabled }
        "400":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
  /api/v1/me/totp/disable:
    post:
      tags: [Auth]
      summary: Turn off two-factor
      description: >-
        Requires the password as well as a current code, so a borrowed
        unlocked session cannot strip the factor. POST rather than DELETE
        because it carries a body.


        If the authenticator is lost, the recovery path is shell access to
        the server: `tradermemos disable-totp --email you@example.com`.
        There are deliberately no recovery codes to store or lose.
      operationId: disableTotp
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [password, code]
              properties:
                password: { type: string }
                code: { type: string }
      responses:
        "204": { description: Disabled }
        "400":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        Session JWT access token **or** personal access token (`tm_pat_…`).
...
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details: {}
...
    Me:
      type: object
      required: [id, email, is_admin, created_at, totp_enabled]
      properties:
        id: { type: string }
        email: { type: string, format: email }
        is_admin: { type: boolean }
        created_at: { type: string, format: date-time }
        totp_enabled:
          type: boolean
          description: Always false today — the column exists but nothing writes it.
    TokenPair:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        is_admin:
          type: boolean
...
security:
  - bearerAuth: []